requirements.txt
fastapi
uvicorn
requests
groq
pydantic
python-dotenv
reportlab